fix interpretSatelliteInfo's reset condition
interpretSatelliteInfo is incremental now, and has to
reset the SatelliteInfo array when a new bunch of messages
is received.

The condition was to look for message 1, but the message count
is constellation-specific (i.e. there is three messages number 1
in a set).

Fix this by resetting only on the first GPS message.