implement a method to wait for the board to be reset
When reset operations like INI or POP are performed, the
boards needs a few seconds to actually be operational again
(around 2 seconds actually). Moreover, it sends a specific
message when it is booted.

This new method waits at most 10s for a proper ASHTEC message
to be available on the port.